Coverity Fix: Improve error handling and edge case handling
- Simplified timestamp casting in "dcmihandler.cpp"
- Added sensor existence check in "sensorhandler.cpp" to prevent
 undefined behavior on missing sensors
- Improved error handling and structured exception management in
 "softoff/mainapp.cpp"
- Zero-filled user name buffer in "user_mgmt.cpp" to ensure null
 termination and prevent buffer overflows by limiting "strncpy" to
 "ipmiMaxUserName - 1"

Fixes NVBug https://nvbugspro.nvidia.com/bug/4921269

+++ dcmihandler.cpp
@@ -1093,7 +1093,7 @@ ipmi::RspType<uint16_t, // current power
     constexpr uint6_t reserved1 = 0;
     constexpr bool measurementActive = true;
     constexpr bool reserved2 = false;
-    auto timestamp = static_cast<uint32_t>(time(nullptr));
+    auto timestamp = time(nullptr);
     return ipmi::responseSuccess(power, power, power, power, timestamp,
                                  samplePeriod, reserved1, measurementActive,
                                  reserved2);
